+++ example/android/app/src/main/kotlin/br/com/examplecube/MainActivity.kt
@@ -1,6 +1,0 @@
-package br.com.examplecube
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}