+++ lib/src/actions/cube_action.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 export 'navigator_action.dart';
 
 ///
@@ -13,4 +15,7 @@ export 'navigator_action.dart';
 /// on 14/02/22
 /// Base to Actions
 
-abstract class CubeAction {}
+abstract class CubeAction {
+  /// When the action is sent he will are executed per this method
+  void execute(BuildContext context);
+}