+++ test/widgets/cube_builder/cube_builder_robot.dart
@@ -3,7 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-class ActionExample extends CubeAction {}
+class ActionExample extends CubeAction {
+  @override
+  void execute(BuildContext context) {}
+}
 
 class CubeExample extends Cube {
   void sendAction(CubeAction action) {